Finish global timer

+++ wwdc19.playground/Sources/GameEvents.swift
@@ -12,7 +12,7 @@ import Foundation
 protocol GameEvents {
     
     var level: Int { get set }
-    var timer: Int {get set}
+    var timer: Double {get set}
     
     func userDidRightChoice(index: Int)
     func userDidWrongChoice()
